[OPR-321]: Removes use of NSThread to check for main queue
