+++ Sources/Core/Shared/ThreadSafety.swift
@@ -107,7 +107,7 @@ public func dispatch_sync<T>(queue: dispatch_queue_t, _ block: () throws -> T) r
 }
 
 internal func dispatch_main_sync<T>(block: () throws -> T) rethrows -> T {
-    guard NSThread.isMainThread() else {
+    guard Queue.isMainQueue else {
         return try dispatch_sync(Queue.Main.queue, block)
     }
     return try block()
